Add error handling

+++ src/controllers/CardController.ts
@@ -8,37 +8,42 @@ import {
   Delete,
   JsonController,
   QueryParams,
-} from "routing-controllers";
-import { Card } from "../entities/Card";
-import { CardService } from "../services/CardService";
+} from 'routing-controllers';
+import { Card } from '../entities/Card';
+import EntityNotFoundError from '../errors/EntityNotFoundError';
+import { CardService } from '../services/CardService';
 
-@JsonController("/cards")
+@JsonController('/cards')
 export class CardController {
   private cardService = new CardService();
 
-  @Get("/")
-  getAll(@QueryParams() query: Card) {
+  @Get('/')
+  public async getAll(@QueryParams() query: Card) {
     return this.cardService.find(query);
   }
 
-  @Get("/:id")
-  getOne(@Param("id") id: string) {
-    return this.cardService.findById(id);
+  @Get('/:id')
+  public async getOne(@Param('id') id: string) {
+    const card = await this.cardService.findById(id);
+    if (!card) throw new EntityNotFoundError();
+    return card;
   }
 
-  @Post("/")
-  post(@Body() card: Card) {
+  @Post('/')
+  public async post(@Body() card: Card) {
     return this.cardService.create(card);
   }
 
-  @Put("/:id")
-  put(@Param("id") id: string, @Body() card: Card) {
-    const safeEntity = {...card,id}
-    return this.cardService.update(id,safeEntity);
+  @Put('/:id')
+  public async put(@Param('id') id: string, @Body() card: Card) {
+    if (!this.cardService.existsById(id)) throw new EntityNotFoundError();
+    const safeEntity = { ...card, id };
+    return this.cardService.update(id, safeEntity);
   }
 
-  @Delete("/:id")
-  remove(@Param("id") id: string) {
-    return this.cardService.deleteById(id)
+  @Delete('/:id')
+  public async remove(@Param('id') id: string) {
+    if (!this.cardService.existsById(id)) throw new EntityNotFoundError();
+    return this.cardService.deleteById(id);
   }
 }
