+++ src/controllers/ClientController.ts
@@ -9,6 +9,7 @@ import {
   QueryParams,
 } from 'routing-controllers';
 import { Client } from '../entities/Client';
+import EntityNotFoundError from '../errors/EntityNotFoundError';
 import { ClientService } from '../services/ClientService';
 
 @JsonController('/clients')
@@ -16,29 +17,32 @@ export class ClientController {
   private clientService = new ClientService();
 
   @Get('/')
-  getAll(@QueryParams() query: Client) {
+  public async getAll(@QueryParams() query: Client) {
     return this.clientService.find(query);
   }
 
   @Get('/:id')
-  getOne(@Param('id') id: number) {
-    return this.clientService.findById(id);
+  public async getOne(@Param('id') id: number) {
+    const client = await this.clientService.findById(id);
+    if (!client) throw new EntityNotFoundError();
   }
 
   @Post('/')
-  async post(@Body() client: Client) {
+  public async post(@Body() client: Client) {
     delete client.id;
     return await this.clientService.create(client);
   }
 
   @Put('/:id')
-  put(@Param('id') id: number, @Body() client: Client) {
+  public async put(@Param('id') id: number, @Body() client: Client) {
+    if (!this.clientService.existsById(id)) throw new EntityNotFoundError();
     const safeEntity = { ...client, id };
     return this.clientService.update(id, safeEntity);
   }
 
   @Delete('/:id')
-  remove(@Param('id') id: number) {
+  public async remove(@Param('id') id: number) {
+    if (!this.clientService.existsById(id)) throw new EntityNotFoundError();
     return this.clientService.deleteById(id);
   }
 }
